packages/yai-loguru-sls-sink/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yai-loguru-sls-sink"
version = "0.1.0"
description = "An Aliyun SLS sink plugin for yai-loguru."
authors = [
    { name="YAI-Nexus", email="contact@yai-nexus.com" },
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Logging",
]
dependencies = [
    "loguru>=0.7.0",
    "aliyun-log-python-sdk>=0.7.8",
    "python-dotenv>=1.0.0",
]

[tool.uv.sources]
yai-loguru = { workspace = true }

[project.entry-points."yai_loguru.plugins"]
sls_sink = "yai_loguru_sls_sink:SlsSinkPlugin"

[tool.hatch.build.targets.wheel]
packages = ["src/yai_loguru_sls_sink"]

[tool.hatch.build.targets.wheel.force-include]
"src/yai_loguru_sls_sink/py.typed" = "yai_loguru_sls_sink/py.typed"